requirements.txt
# Core processing - stable versions
pytube>=15.0.0,<16.0.0
yt-dlp>=2023.12.30,<2024.1.0
moviepy>=1.0.3,<2.0.0
ffmpeg-python>=0.2.0,<1.0.0
pydub>=0.25.1,<1.0.0

# AI and machine learning - compatible versions
openai-whisper>=20231117
torch>=2.0.0,<3.0.0
torchaudio>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
speechbrain>=0.5.0,<1.0.0
transformers>=4.30.0,<5.0.0

# Numerical computing - fixed compatibility
numpy>=1.24.0,<2.1.0  # Fix for Whisper/numba compatibility
scipy>=1.10.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
pandas>=2.0.0,<3.0.0

# Audio processing
librosa>=0.10.0,<1.0.0
noisereduce>=3.0.0,<4.0.0

# NLP and content analysis
spacy>=3.6.0,<4.0.0
nltk>=3.8.0,<4.0.0
keybert>=0.8.0,<1.0.0
textstat>=0.7.0,<1.0.0

# Computer vision (optional)
opencv-python>=4.8.0,<5.0.0

# Web and API
flask>=2.3.0,<3.0.0
fastapi>=0.100.0,<1.0.0
uvicorn>=0.23.0,<1.0.0
requests>=2.31.0,<3.0.0

# Security and validation
cryptography>=41.0.0,<42.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0

# GUI
pillow>=10.0.0,<11.0.0
tkinter-tooltip>=2.2.0,<3.0.0

# Development and testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-mock>=3.10.0,<4.0.0

# Utilities
tqdm>=4.65.0,<5.0.0
click>=8.1.0,<9.0.0
colorama>=0.4.6,<1.0.0
python-dateutil>=2.8.0,<3.0.0
psutil>=5.9.0,<6.0.0

# Optional dependencies for specific features
# Install with: pip install -r requirements.txt -r requirements-optional.txt
# google-api-python-client>=2.95.0  # For Google services
# dropbox>=11.36.0                  # For Dropbox integration
# azure-storage-blob>=12.17.0       # For Azure storage